// src/enums/actions.ts
// Enums for context actions
export enum Actions {
  SET_BG_COLOR_VALUE = 'SET_BG_COLOR_VALUE',
  SET_SELECTED_ASSET = 'SET_SELECTED_ASSET',
  SET_CONNECTED_NETWORK = 'SET_CONNECTED_NETWORK',
  SET_SELECTED_TAB = 'SET_SELECTED_TAB',
  SET_ASSET_PRICE = 'SET_ASSET_PRICE',
  SET_PRICE_RANGE = 'SET_PRICE_RANGE',
  SET_PRINCIPAL_TIME_TO_EXPIRY = 'SET_PRINCIPAL_TIME_TO_EXPIRY',
  SET_CRUIZE_CONTRACT = 'SET_CRUIZE_CONTRACT',
  SET_SELECTED_ASSET_CONTRACT = 'SET_SELECTED_ASSET_CONTRACT',
  SET_SELCTED_ASSET_APPROVED = 'SET_TOKEN_APPROVED',
  SET_BALANCES = 'SET_BALANCES',
  SET_TRANSACTION_DETAILS = 'SET_TRANSACTION_DETAILS',
  SET_MINT_TOKEN_CONTRACT = 'SET_MINT_TOKEN_CONTRACT',
  SET_TRANSACTION_DATA = 'SET_TRANSACTION_DATA',
  SET_USER_INPUT_VALUE = 'SET_USER_INPUT_VALUE',
  SET_APP_ERROR = 'SET_APP_ERROR',
  SET_LOCKED_ASSET = 'SET_LOCKED_ASSET',
  SET_CURRENT_DEPOSIT = 'SET_CURRENT_DEPOSIT',
  SET_WITHDRAW_TYPE = 'SET_WITHDRAW_TYPE',
  SET_YIELD_INFO_DATA = 'SET_YIELD_INFO_DATA',
  SET_APPROVE_TOKEN_MODAL = 'SET_APPROVE_TOKEN_MODAL',
  SET_ASSET_APYS = 'SET_ASSET_APYS',
  SET_RAMSES_VAULT_CONTRACT = 'SET_RAMSES_VAULT_CONTRACT',
  SET_RAMSES_VAULT_PAIR = 'SET_RAMSES_VAULT_PAIR'
}